refactor DB handling into own SqliteService
