+++ src/app/shared-services/shared-services.module.ts
@@ -3,6 +3,7 @@ import { HttpClientModule } from '@angular/common/http'
 import { NgModule } from '@angular/core'
 import { BackgroundGeolocation } from '@ionic-native/background-geolocation/ngx'
 import { LocalNotifications } from '@ionic-native/local-notifications/ngx'
+import { SqliteService } from './db/sqlite.service'
 import { LocationService } from './location.service'
 import { TrajectoryService } from './trajectory.service'
 
@@ -15,6 +16,7 @@ import { TrajectoryService } from './trajectory.service'
     // providers
     LocationService,
     TrajectoryService,
+    SqliteService,
   ],
   imports: [CommonModule, HttpClientModule],
 })
